backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-decouple==3.8
Pillow==10.1.0
openai==1.3.5
django-extensions==3.2.3
watchdog==3.0.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

# AWS Integration
boto3==1.34.20
django-storages==1.14.2
django-environ==0.11.2
